Create 2)Read and Print an Integer value in Java
The given task is to take an integer as input from the user and print that integer in Java language. 

In below program, the syntax and procedures to take the integer as input from the user is shown in Java language. Steps:

The user enters an integer value when asked.
This value is taken from the user with the help of nextInt() method of Scanner Class. The nextInt() method, in Java, reads the next integer value from the console into the specified variable.